Vllm/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
pydantic>=2.5.0
# Note: vLLM is already included in the base rocm/vllm-dev:nightly image